requirements.txt
pysmb
tkinterdnd2
pywin32
Pillow
cryptography
